+++ module/Substitution/src/Substitution/Service/Substitution/EcoleDoctorale/EcoleDoctoraleSubstitutionService.php
@@ -3,6 +3,7 @@
 namespace Substitution\Service\Substitution\EcoleDoctorale;
 
 use Application\Service\BaseService;
+use Substitution\Constants;
 use Substitution\Entity\Db\SubstitutionAwareEntityInterface;
 use Substitution\Service\Substitution\SpecificSubstitutionAbstractService;
 use Substitution\Service\Substitution\Structure\StructureConcreteSubstitutionServiceTrait;
@@ -11,15 +12,12 @@ class EcoleDoctoraleSubstitutionService extends SpecificSubstitutionAbstractServ
 {
     use StructureConcreteSubstitutionServiceTrait;
 
+    protected string $type = Constants::TYPE_ecole_doct;
+
     /**
      * @var \Structure\Service\EcoleDoctorale\EcoleDoctoraleService
      */
     protected BaseService $entityService;
-
-    public function __construct()
-    {
-        $this->type = 'ecole_doct';
-    }
 
     public function saveEntity(SubstitutionAwareEntityInterface $entity): void
     {