+++ module/Substitution/src/Substitution/Service/Substitution/UniteRecherche/UniteRechercheSubstitutionService.php
@@ -3,6 +3,7 @@
 namespace Substitution\Service\Substitution\UniteRecherche;
 
 use Application\Service\BaseService;
+use Substitution\Constants;
 use Substitution\Entity\Db\SubstitutionAwareEntityInterface;
 use Substitution\Service\Substitution\SpecificSubstitutionAbstractService;
 use Substitution\Service\Substitution\Structure\StructureConcreteSubstitutionServiceTrait;
@@ -11,15 +12,12 @@ class UniteRechercheSubstitutionService extends SpecificSubstitutionAbstractServ
 {
     use StructureConcreteSubstitutionServiceTrait;
 
+    protected string $type = Constants::TYPE_unite_rech;
+
     /**
      * @var \Structure\Service\UniteRecherche\UniteRechercheService
      */
     protected BaseService $entityService;
-
-    public function __construct()
-    {
-        $this->type = 'unite_rech';
-    }
 
     public function saveEntity(SubstitutionAwareEntityInterface $entity): void
     {