+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Journal des modifications
 - Stockage de fichier S3/Filesystem : plus de copie sur disque dans le cas d'une demande d'un fichier issu du Filesystem.
 - Inclusion au script de purge des fichiers temporaires ceux générés par le FichierStorageService.
 - Module Substitutions : amélioration de fonctions pgsql du moteur de substitutions.
+- Module Substitutions : possibilité de lancer la création d'une substitution depuis la page des doublons.
 - Structures : amélioration visuelle sur la page listant les structures (Etablissement, ED, UR).
 - Listing/recherche d'individus : lien vers la substitution si individu substituant ; nouveau filtre 'Historisés inclus' décoché par défaut ; cosmétique.  
 - [FIX] Module Substitutions : plantage de la page consacrée aux doublons (substitutions possibles).